Moved Widget code from WidgetActivity to SurfaceActivity
