+++ app/src/main/AndroidManifest.xml
@@ -1196,21 +1196,8 @@
                 android:value=".ExampleOverviewActivity" />
         </activity>
         <activity
-            android:name=".examples.WidgetActivity"
-            android:description="@string/description_widget"
-            android:exported="true"
-            android:label="@string/activity_widget">
-            <meta-data
-                android:name="@string/category"
-                android:value="@string/category_widget" />
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value=".ExampleOverviewActivity" />
-        </activity>
-        <activity
             android:name=".TestMapActivity"
             android:exported="true" />
-
         <meta-data
             android:name="com.mapbox.maps.ThreadChecker"
             android:value="true" />
