+++ app/src/main/java/com/mapbox/maps/testapp/examples/SurfaceActivity.kt
@@ -1,10 +1,17 @@
 package com.mapbox.maps.testapp.examples
 
+import android.animation.ValueAnimator
 import android.annotation.SuppressLint
+import android.content.Context
+import android.graphics.BitmapFactory
 import android.os.Bundle
 import android.view.SurfaceHolder
+import android.view.animation.Animation
 import androidx.appcompat.app.AppCompatActivity
 import com.mapbox.maps.*
+import com.mapbox.maps.renderer.widget.BitmapWidget
+import com.mapbox.maps.renderer.widget.Widget
+import com.mapbox.maps.renderer.widget.WidgetPosition
 import com.mapbox.maps.testapp.databinding.ActivitySurfaceBinding
 
 /**
@@ -14,6 +21,11 @@ class SurfaceActivity : AppCompatActivity(), SurfaceHolder.Callback {
 
   private lateinit var surfaceHolder: SurfaceHolder
   private lateinit var mapSurface: MapSurface
+  private lateinit var widget: Widget
+  private var animator = ValueAnimator.ofFloat(0f, 1f).also {
+    it.duration = ANIMATION_DURATION
+    it.repeatCount = ValueAnimator.INFINITE
+  }
 
   @SuppressLint("ClickableViewAccessibility")
   public override fun onCreate(savedInstanceState: Bundle?) {
@@ -24,9 +36,14 @@ class SurfaceActivity : AppCompatActivity(), SurfaceHolder.Callback {
     // Setup map surface
     surfaceHolder = binding.surface.holder
     surfaceHolder.addCallback(this)
+    val mapOptions = MapInitOptions.getDefaultMapOptions(this).toBuilder().also {
+      it.contextMode(ContextMode.SHARED)
+    }.build()
+    val mapInitOptions = MapInitOptions(this, mapOptions = mapOptions)
     mapSurface = MapSurface(
       this,
-      surfaceHolder.surface
+      surfaceHolder.surface,
+      mapInitOptions,
     )
 
     // Load a map style
@@ -35,6 +52,19 @@ class SurfaceActivity : AppCompatActivity(), SurfaceHolder.Callback {
     // Touch handling (verify plugin integration)
     binding.surface.setOnTouchListener { _, event -> mapSurface.onTouchEvent(event) }
     binding.surface.setOnGenericMotionListener { _, event -> mapSurface.onGenericMotionEvent(event) }
+
+    val widgetPosition = WidgetPosition
+      .Builder()
+      .setHorizontalAlignment(WidgetPosition.Horizontal.CENTER)
+      .setVerticalAlignment(WidgetPosition.Vertical.CENTER)
+      .build()
+    widget = LogoWidget(this, widgetPosition)
+    mapSurface.addWidget(widget)
+    animator.addUpdateListener {
+      val angle = (it.animatedFraction * 360f) % 360f
+      widget.setRotation(angle)
+    }
+    animator.start()
   }
 
   override fun surfaceCreated(holder: SurfaceHolder) {
@@ -64,7 +94,23 @@ class SurfaceActivity : AppCompatActivity(), SurfaceHolder.Callback {
     mapSurface.onDestroy()
   }
 
+  override fun onResume() {
+    super.onResume()
+    animator.resume()
+  }
+
+  override fun onPause() {
+    super.onPause()
+    animator.pause()
+  }
+
   companion object {
     const val TAG = "SurfaceActivity"
+    const val ANIMATION_DURATION = 10000L
   }
+
+  private class LogoWidget constructor(context: Context, position: WidgetPosition) : BitmapWidget(
+    BitmapFactory.decodeResource(context.resources, R.drawable.mapbox_logo_icon),
+    position,
+  )
 }
